+++ src/nodejs/client.ts
@@ -6,9 +6,12 @@ import {
   ClientStatus,
   Connection,
   ConnectOptions,
+  DataQueryOptions,
+  DataQueryResponse,
   getEndpoints,
   ListenerCallback,
   NstrumentaClientBase,
+  StorageUploadParameters,
   SubscriptionCallback,
 } from '../shared';
 import {
@@ -16,15 +19,22 @@ import {
   makeBusMessageFromBuffer,
   makeBusMessageFromJsonObject,
 } from '../shared/lib/busMessage';
+import { LogConfig } from './server';
 
 const endpoints = process.env.NSTRUMENTA_LOCAL ? getEndpoints('local') : getEndpoints('prod');
+
+type Reconnection = {
+  hasVerified: boolean;
+  attempts: number;
+  timeout: ReturnType<typeof setTimeout> | null;
+};
 
 export class NstrumentaClient implements NstrumentaClientBase {
   private ws: WebSocket | null = null;
   private apiKey?: string;
   private listeners: Map<string, Array<ListenerCallback>>;
   private subscriptions: Map<string, SubscriptionCallback[]>;
-  private reconnection = { hasVerified: false, attempts: 0 };
+  private reconnection: Reconnection = { hasVerified: false, attempts: 0, timeout: null };
   private messageBuffer: Array<ArrayBufferLike>;
   private datalogs: Map<string, Array<string>>;
   public clientId: string | null = null;
@@ -46,6 +56,10 @@ export class NstrumentaClient implements NstrumentaClientBase {
     this.subscriptions.clear();
     this.datalogs.clear();
     this.messageBuffer = [];
+    if (this.reconnection.timeout) {
+      clearTimeout(this.reconnection.timeout);
+      this.reconnection.timeout = null;
+    }
     this.ws?.removeAllListeners();
     this.ws?.close();
     return;
@@ -93,7 +107,8 @@ export class NstrumentaClient implements NstrumentaClientBase {
         this.subscriptions.clear();
         // reconnect on close
         if (this.reconnection.hasVerified) {
-          setTimeout(() => {
+          this.reconnection.timeout = setTimeout(() => {
+            this.reconnection.timeout = null;
             console.log(`attempting to reconnect, attempts: ${this.reconnection.attempts}`);
             this.connect(connectOptions);
           }, this.rollOff(this.reconnection.attempts));
@@ -185,9 +200,9 @@ export class NstrumentaClient implements NstrumentaClientBase {
     }
   }
 
-  public async startLog(name: string, channels: string[]) {
+  public async startLog(name: string, channels: string[], config?: LogConfig) {
     // TODO error on slashes ?
-    this.send('_nstrumenta', { command: 'startLog', name, channels });
+    this.send('_nstrumenta', { command: 'startLog', name, channels, config });
   }
 
   public async finishLog(name: string) {
@@ -218,6 +233,40 @@ class StorageService implements BaseStorageService {
     return download.data;
   }
 
+  async query({
+    filenames,
+    tag: tags,
+    before,
+    after,
+    limit = 1,
+    metadata: metadataOriginal,
+  }: DataQueryOptions): Promise<DataQueryResponse> {
+    const metadata =
+      typeof metadataOriginal === 'string'
+        ? JSON.parse(metadataOriginal)
+        : typeof metadataOriginal === 'object'
+        ? metadataOriginal
+        : {};
+
+    const data = { tags, before, after, limit, filenames, metadata };
+
+    const config: AxiosRequestConfig = {
+      method: 'post',
+      headers: { 'x-api-key': this.apiKey },
+      data,
+    };
+
+    try {
+      const response = await axios(endpoints.QUERY_DATA, config);
+
+      return response.data;
+    } catch (error) {
+      console.log(`Something went wrong: ${(error as Error).message}`);
+
+      return [];
+    }
+  }
+
   async list(type: string): Promise<string[]> {
     let response = await axios(endpoints.LIST_STORAGE_OBJECTS, {
       method: 'post',
@@ -228,13 +277,23 @@ class StorageService implements BaseStorageService {
     return response.data;
   }
 
-  async upload(path: string, data: Blob, meta: Record<string, string>): Promise<void> {
+  async upload({
+    filename,
+    data,
+    meta,
+    dataId: explicitDataId,
+  }: StorageUploadParameters): Promise<void> {
+    console.log('uploading', filename, meta, explicitDataId, data.size);
     const size = data.size;
     let url;
-    const res = await axios(endpoints.GENERATE_DATA_ID, {
-      headers: { 'x-api-key': this.apiKey!, method: 'post' },
-    });
-    const dataId = res.data.dataId;
+    let dataId = explicitDataId;
+
+    if (!dataId) {
+      const res = await axios(endpoints.GENERATE_DATA_ID, {
+        headers: { 'x-api-key': this.apiKey!, method: 'post' },
+      });
+      dataId = res.data.dataId;
+    }
     const config: AxiosRequestConfig = {
       method: 'post',
       headers: {
@@ -242,8 +301,8 @@ class StorageService implements BaseStorageService {
         'Content-Type': 'application/json',
       },
       data: {
-        name: path,
-        dataId,
+        name: filename,
+        dataId: dataId,
         size,
         metadata: meta,
       },
@@ -253,7 +312,7 @@ class StorageService implements BaseStorageService {
 
     url = response.data?.uploadUrl;
     if (!url) {
-      console.warn(`no upload url returned, can't upload ${path}`);
+      console.warn(`no upload url returned, can't upload ${filename}`);
       console.log(response.data);
       return;
     }