+++ src/nodejs/server.ts
@@ -1,3 +1,4 @@
+import { Mcap0Types, Mcap0Writer as McapWriter } from '@mcap/core';
 import axios from 'axios';
 import { ChildProcess } from 'child_process';
 import { randomUUID } from 'crypto';
@@ -5,10 +6,11 @@ import express from 'express';
 import { createWriteStream } from 'fs';
 import fs from 'fs/promises';
 import serveIndex from 'serve-index';
-import { Writable } from 'stream';
+import { Readable, Transform, Writable } from 'stream';
 import { WebSocket, WebSocketServer } from 'ws';
 import { asyncSpawn, createLogger, getNstDir, resolveApiKey } from '../cli/utils';
 import { DEFAULT_HOST_PORT, getEndpoints } from '../shared';
+
 import {
   BusMessageType,
   deserializeWireMessage,
@@ -18,8 +20,18 @@ import {
 import { verifyToken } from '../shared/lib/sessionToken';
 import { NstrumentaClient } from './client';
 import { start as startVideoServer } from './video/examples/server-demo/src/main';
+import { FileHandleWritable } from './fileHandleWriteable';
 import WritableStream = NodeJS.WritableStream;
+
 const endpoints = process.env.NSTRUMENTA_LOCAL ? getEndpoints('local') : getEndpoints('prod');
+
+const createPrefixTransform = (prefix: string) =>
+  new Transform({
+    transform(chunk, encoding, callback) {
+      const str = `${prefix} ${chunk}`;
+      callback(null, str);
+    },
+  });
 
 const logger = createLogger();
 
@@ -37,6 +49,7 @@ export interface CommandRunModuleData {
   module: string;
   actionId: string;
   args?: string[];
+  version?: string;
 }
 
 export interface CommandStopModuleData {
@@ -44,6 +57,14 @@ export interface CommandStopModuleData {
 }
 
 export type AgentActionStatus = 'pending' | 'complete';
+
+export interface LogConfig {
+  header: Mcap0Types.Header;
+  channels: Array<{
+    schema: { title: string; type: 'object'; properties: Record<string, unknown> };
+    channel: Omit<Mcap0Types.Channel, 'id' | 'schemaId' | 'metadata'>;
+  }>;
+}
 
 export type BackplaneCommand =
   | {
@@ -69,6 +90,21 @@ export interface NstrumentaServerOptions {
   allowUnverifiedConnection?: boolean;
 }
 
+export type DataLog =
+  | {
+      type: 'stream';
+      localPath: string;
+      channels: string[];
+      stream: WritableStream;
+    }
+  | {
+      type: 'mcap';
+      localPath: string;
+      channels: string[];
+      channelIds: Map<string, number>;
+      mcapWriter: McapWriter;
+    };
+
 export class NstrumentaServer {
   options: NstrumentaServerOptions;
   backplaneClient?: NstrumentaClient;
@@ -79,8 +115,7 @@ export class NstrumentaServer {
   children: Map<string, ChildProcess> = new Map();
   cwd = process.cwd();
   logStreams?: Writable[];
-  dataLogs: Map<string, { localPath: string; channels: string[]; stream: WritableStream }> =
-    new Map();
+  dataLogs: Map<string, DataLog> = new Map();
   status: ServerStatus;
 
   constructor(options: NstrumentaServerOptions) {
@@ -145,7 +180,9 @@ export class NstrumentaServer {
           actionsCollectionPath,
         }: { backplaneUrl: string; agentId: string; actionsCollectionPath: string } = response.data;
 
-        logger.setPrefix(`[agent ${this.options.tag || `${agentId.substring(0, 5)}...`}]`);
+        logger.setPrefix(
+          `(${new Date()})[agent ${this.options.tag || `${agentId.substring(0, 5)}...`}]`
+        );
         this.status.agentId = agentId;
         if (backplaneUrl) {
           this.backplaneClient.addListener('open', () => {
@@ -165,7 +202,7 @@ export class NstrumentaServer {
                     return;
                   }
                   const {
-                    data: { module: moduleName, args },
+                    data: { module: moduleName, args, version },
                   } = message;
 
                   const logPath = `${await getNstDir(this.cwd)}/${moduleName}-${actionId}.txt`;
@@ -176,22 +213,28 @@ export class NstrumentaServer {
                     next();
                   };
 
-                  const process = await asyncSpawn(
+                  console.log('running module', moduleName, version);
+                  const childProcess = await asyncSpawn(
                     'nstrumenta',
                     [
                       'module',
                       'run',
                       `--name=${moduleName}`,
+                      version ? `--module-version=${version}` : '',
                       '--non-interactive',
                       '--',
                       ...(args ? args : []),
                     ],
-                    undefined,
+                    { stdio: 'pipe' },
                     undefined,
                     [stream, backplaneStream]
                   );
-                  this.children.set(String(process.pid), process);
-                  process.on('disconnect', () => this.children.delete(String(process.pid)));
+                  this.children.set(String(childProcess.pid), childProcess);
+                  childProcess.on('disconnect', () =>
+                    this.children.delete(String(childProcess.pid))
+                  );
+                  const transform = createPrefixTransform(`${moduleName}/${actionId}`);
+                  childProcess.stdout = childProcess.stdout?.pipe(transform) as Readable;
                   break;
                 case 'stopModule':
                   break;
@@ -352,10 +395,10 @@ export class NstrumentaServer {
         // commands from clients
 
         if (contents?.command == 'startLog') {
-          const { channels, name } = contents;
-          logger.log(`[nstrumenta] <startLog> ${channels}`);
+          const { channels, name, config } = contents;
+          logger.log(`[nstrumenta] <startLog> ${channels} ${JSON.stringify(config)}`);
           try {
-            await this.startLog(name, channels);
+            await this.startLog(name, channels, config);
           } catch (error) {
             logger.log((error as Error).message);
           }
@@ -386,12 +429,41 @@ export class NstrumentaServer {
           }
         });
 
-        this.dataLogs.forEach((dataLog) => {
+        this.dataLogs.forEach(async (dataLog) => {
           if (dataLog.channels.includes(channel)) {
-            if (busMessageType === BusMessageType.Json) {
-              dataLog.stream.write(JSON.stringify({ channel, contents }) + '\n');
-            } else {
-              dataLog.stream.write(busMessage);
+            switch (dataLog.type) {
+              case 'stream':
+                {
+                  if (busMessageType === BusMessageType.Json) {
+                    dataLog.stream.write(JSON.stringify({ channel, contents }) + '\n');
+                  } else {
+                    dataLog.stream.write(busMessage);
+                  }
+                }
+                break;
+              case 'mcap': {
+                const channelId = dataLog.channelIds.get(channel);
+                if (channelId === undefined)
+                  throw new Error(
+                    `attempting to log unregistered mcap channel ${channel} - be sure to add channel to LogConfig`
+                  );
+
+                const logTime =
+                  contents.timestamp !== undefined
+                    ? contents.timestamp.sec !== undefined
+                      ? BigInt(contents.timestamp.sec) * BigInt(1_000_000_000) +
+                        BigInt(contents.timestamp.nsec)
+                      : BigInt(contents.timestamp) * BigInt(1_000_000)
+                    : BigInt(Date.now()) * BigInt(1_000_000);
+
+                await dataLog.mcapWriter.addMessage({
+                  channelId,
+                  sequence: 0,
+                  publishTime: logTime,
+                  logTime: logTime,
+                  data: Buffer.from(JSON.stringify(contents)),
+                });
+              }
             }
           }
         });
@@ -426,24 +498,66 @@ export class NstrumentaServer {
     });
   }
 
-  public async startLog(name: string, channels: string[]) {
+  public async startLog(name: string, channels: string[], config?: LogConfig) {
     logger.log('[server] <startLog>', { channels });
     const nstDir = await getNstDir(this.cwd);
     await fs.mkdir(`${nstDir}/data`, { recursive: true }); // mkdir in case data dir doesn't yet exist
     const localPath = `${nstDir}/data/${randomUUID()}`;
     logger.log({ name, cwdNstDir: nstDir, localPath });
-    const stream = await createWriteStream(localPath);
-    this.dataLogs.set(name, { localPath, channels, stream });
+    if (config) {
+      const type = 'mcap';
+      const fileHandle = await fs.open(localPath, 'w');
+      const fileHandleWritable = new FileHandleWritable(fileHandle);
+      const mcapWriter = new McapWriter({
+        writable: fileHandleWritable,
+        useStatistics: true,
+        useChunks: true,
+        useChunkIndex: true,
+      });
+      await mcapWriter.start(config.header);
+
+      const channelIds = new Map();
+      await Promise.all(
+        config.channels.map(async (c) => {
+          const schemaId = await mcapWriter.registerSchema({
+            name: c.schema.title,
+            encoding: 'jsonschema',
+            data: Buffer.from(JSON.stringify(c.schema)),
+          });
+
+          const channelId = await mcapWriter.registerChannel({
+            schemaId,
+            metadata: new Map(),
+            ...c.channel,
+          });
+          return channelIds.set(c.channel.topic, channelId);
+        })
+      );
+
+      this.dataLogs.set(name, { type, localPath, channels, channelIds, mcapWriter });
+    } else {
+      const type = 'stream';
+      const stream = await createWriteStream(localPath);
+      this.dataLogs.set(name, { type, localPath, channels, stream });
+    }
 
     return this.dataLogs;
   }
 
   public async finishLog(name: string) {
     logger.log(`<finishLog>: ${name}`);
-    if (!this.dataLogs.has(name)) throw new Error('log name not found');
-    const { localPath, stream } = this.dataLogs.get(name)!;
-    logger.log(`Ending stream: ${localPath}`);
-    stream.end();
+    const dataLog = this.dataLogs.get(name);
+    if (!dataLog) throw new Error('log name not found');
+    const { localPath } = dataLog;
+    logger.log(`Finishing log: ${dataLog?.localPath}`);
+    switch (dataLog?.type) {
+      case 'stream':
+        dataLog.stream.end();
+        break;
+      case 'mcap':
+        dataLog.mcapWriter.end();
+    }
+
     this.dataLogs.delete(name);
 
     try {
@@ -507,7 +621,6 @@ const getStorageUploadIntervalStream = (agentId: string) => {
   const stream = new Writable();
   const buffer = { current: '' };
   stream._write = (chunk, enc, next) => {
-    console.log('***', chunk);
     buffer.current += chunk;
     next();
   };